Moved json codec logic to converter. Now converters implement json codec and are always generated
